a temporary fix that converts indices to double before passing to indexmax. indexmax needs to be replaced by indexmaxi which has better handling of Iexpand vectors and which uses uint64 indexing which should improve performance relative to using doubles

+++ mdputils/mdpsolve_Inf.m
@@ -286,7 +286,7 @@ function [vnew,xnew] = valmax(v)
   end
   vnew=R+d.*vnew;
   if Xindexed
-    [vnew,xnew]=indexmax(vnew,Ix,ns);  % use mex version for greater speed
+    [vnew,xnew]=indexmax(vnew,double(Ix),double(ns));  % use mex version for greater speed
   else
     [vnew,xnew]=max(reshape(vnew,ns,na),[],2);
   end
